requirements.txt
datetime
requests
python-dateutil
aiohttp
python-dotenv
psycopg2-binary
